bump golang to 1.21.10 and some deps
Bump golang to 1.21.10 and x/net to v0.23.0, x/crypto to v0.17.0 and
google.golang.org/protobuf to v1.33.0.

test/go.mod is bumped from 1.21.6 to 1.21.9. .9 is needed for the
stdlib vulnerability fixes, and some feature is requiring at least .6
to be mentioned, so it cannot be set to just "go 1.21" as we usually
have.

https://osv.dev/GHSA-4v7x-pqxf-cx7m
https://osv.dev/GO-2024-2687
https://osv.dev/GHSA-45x7-px36-x8w8
https://osv.dev/GHSA-8r3f-844c-mc37

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>