Rework the ramdisk API bits
The word "ramdisk" may be ambiguous for users: it may also mean the
ramdisk of the final OS or the Ironic's "ramdisk deploy" feature.

Group all IPA-related bits under a DeployRamdisk structure.

Add a way to disable the downloader (requested in #40). Rename the
downloader container to also use "ramdisk" instead of "ipa".

Closes: #40
Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>