Rework naming of references in the API
As discussed in
https://github.com/metal3-io/ironic-standalone-operator/issues/91, lets
use simpler reference names and reserve the SomethingRef format for when
the field can refer to different kinds or at least namespaces.

This is quite a breaking change but I don't see a lot of value in going
through deprecation since it's going to significantly complicate the
code nobody should be using in production yet.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>