Rework error handling in the controller
Currently, many error conditions are not reported on the Ironic object,
getting it stuck in e.g. "new version detected". This change makes
certain errors be reported in the status conditions. This includes
inability to find the database or the API secret, as well as validation
failures on the API secret.

The database link does pose a UX problem right now: it's not possible to
change it once it's set. I will leave this as a known issue for now.

Closes: #105
Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>