Harden the validation around the ironic version
Currently, the only validation exists on the webhook level. When the
webhook is disabled or the resource is upgraded from an older version,
the reconciliation may start with a wrong version and may fail or even
panic later on.

To avoid it, introduce validation on both CLI and controller levels.
Add a new Version structure that can be used later for version-specific
logic.

Note that this change has an upgrade impact: an Ironic with a no longer
supported version now must be updated before an upgrade.

Closes: #125
Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>