Fix probes for httpd when the downloader is disabled
The current probes rely on /images path which is created by the
downloader, so the probe fails when it's disabled.

Even worse, we're relying on an HTTP redirect from /images to /images/ being
interpreted as success by cURL. Adding -L uncovers the issue: /images/
returns HTTP 404. So, use -L to avoid false positives and switch to
a known path that the downloader creates.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>