add dependabot workflow
In addition to dependabot config that allows dependabot to create PRs
for bumps, for Go ecosystem we also need a dependabot workflow that
after bumps also adds commit with result of "make modules", and
"make generate" to each dependabot commit.

I forgot to add this when adding dependabot config for the Go.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>