Run MariaDB database upgrades as Kubernetes Jobs
Modify the way database upgrades work when MariaDB is used instead of
SQLite. Previously, we would always run the schema upgrade inside
the Ironic start-up script (and completely ignore online migrations).

In an HA deployment, this will cause several Ironic replicas to race
to update the database with unpredictable consequences. Not running
online migrations means that existing objects may get stuck on their
old versions, preventing the next upgrade.

Now separate jobs are started for the schema upgrade and the online
migrations before and after the actual daemon set update. The same
process is used for initializing a new database (but not with SQLite).

The procedure will only work when the updated version of Ironic is
29.0 or newer since the upgrade scripts were not present in 28.0.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>