Use MariaDB operator in the functional tests
To replace the deprecated mariadb-image and IronicDatabase API, this
change introduces mariadb-operator into the functional testing.

A single instance is configured during the preparation phase. To ensure
that tests are isolated, a database and a user are created for each test.

Unfortunately, the dependency on mariadb-operator goes the Go version in
tests to be raised and quite a few transient dependencies to be added.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>