add link checker workflow
Add link checker workflow, so it gets replicated to soon-to-be-created
release branches as well. Schedule runs on 15th of each month.

The one link to be fixed is auto-generated by a job, so it needs to be
fixed each time in the generation jbo, unless we find a way to
generate it with more fresh link.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>