Move the JSON RPC port from 8089 to 6189
The port 8089 conflicts with kubernetes-nmstate, and is generally quite
risky because of the 808* range. We have chosen a new port from the same
618* range we use for httpd.

A quick internet search revealed no known users of port 6189. A query to
Gemini revealed a malware that was used around 2010 but has all but
disappeared by now.

A new API is added to modify the port in case of future conflicts.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>